requirements.txt
tqdm
langchain
javascript
setuptools
openai
chardet
cchardet
chromadb
tiktoken
requests
setuptools
gymnasium
psutil
minecraft_launcher_lib
hydra-core
langchain-openai
langchain-community
langchain-chroma
langchain-deepseek-official
langchain-ollama